HTML/DOC rendering of fmt-related; remove cross-reference rendering in fmt-related if redundant: https://github.com/metanorma/metanorma-ribose/issues/428
