+++ lib/isodoc/presentation_xml_convert.rb
@@ -2,6 +2,7 @@ require_relative "presentation_function/block"
 require_relative "presentation_function/list"
 require_relative "presentation_function/reqt"
 require_relative "presentation_function/concepts"
+require_relative "presentation_function/designations"
 require_relative "presentation_function/terms"
 require_relative "presentation_function/xrefs"
 require_relative "presentation_function/erefs"
