+++ spec/isodoc/terms_spec.rb
@@ -180,449 +180,449 @@ RSpec.describe IsoDoc do
     INPUT
 
     presxml = <<~PRESXML
-       <iso-standard xmlns="http://riboseinc.com/isoxml" type="presentation">
-          <preface>
-             <clause type="toc" id="_" displayorder="1">
-                <fmt-title id="_" depth="1">Table of contents</fmt-title>
-             </clause>
-          </preface>
-          <sections>
-             <terms id="terms_and_definitions" obligation="normative" displayorder="2">
-                <title id="_">Terms and Definitions</title>
-                <fmt-title id="_" depth="1">
-                   <span class="fmt-caption-label">
-                      <semx element="autonum" source="terms_and_definitions">1</semx>
-                      <span class="fmt-autonum-delim">.</span>
-                   </span>
-                   <span class="fmt-caption-delim">
-                      <tab/>
-                   </span>
-                   <semx element="title" source="_">Terms and Definitions</semx>
-                </fmt-title>
-                <fmt-xref-label>
-                   <span class="fmt-element-name">Clause</span>
-                   <semx element="autonum" source="terms_and_definitions">1</semx>
-                </fmt-xref-label>
-                <p>For the purposes of this document, the following terms and definitions apply.</p>
-                <term id="paddy1">
-                   <fmt-name id="_">
-                      <span class="fmt-caption-label">
-                         <semx element="autonum" source="terms_and_definitions">1</semx>
-                         <span class="fmt-autonum-delim">.</span>
-                         <semx element="autonum" source="paddy1">1</semx>
-                         <span class="fmt-autonum-delim">.</span>
-                      </span>
-                   </fmt-name>
-                   <fmt-xref-label>
-                      <span class="fmt-element-name">Clause</span>
-                      <semx element="autonum" source="terms_and_definitions">1</semx>
-                      <span class="fmt-autonum-delim">.</span>
-                      <semx element="autonum" source="paddy1">1</semx>
-                   </fmt-xref-label>
-                   <preferred id="_">
-                      <expression>
-                         <name>paddy</name>
-                      </expression>
-                      <field-of-application id="_">in agriculture</field-of-application>
-                      <usage-info id="_">dated</usage-info>
-                      <source status="modified" id="_">
-                         <origin bibitemid="ISO7301" type="inline" citeas="ISO 7301:2011">
-                            <locality type="clause">
-                               <referenceFrom>3.1</referenceFrom>
-                            </locality>
-                         </origin>
-                         <modification id="_">
-                            <p id="_">The term "cargo rice" is shown as deprecated, and Note 1 to entry is not included here</p>
-                         </modification>
-                      </source>
-                   </preferred>
-                   <fmt-preferred>
-                      <p>
-                         <semx element="preferred" source="_">
-                            <strong>paddy</strong>
-                            <span class="fmt-designation-field">
-                               , &lt;
-                               <semx element="field-of-application" source="_">in agriculture</semx>
-                               ,
-                               <semx element="usage-info" source="_">dated</semx>
-                               &gt;
-                            </span>
-                         </semx>
-                         [SOURCE:
-                         <semx element="source" source="_">
-                            <origin bibitemid="ISO7301" type="inline" citeas="ISO 7301:2011" id="_">
-                               <locality type="clause">
-                                  <referenceFrom>3.1</referenceFrom>
-                               </locality>
-                            </origin>
-                            <semx element="origin" source="_">
-                               <fmt-origin bibitemid="ISO7301" type="inline" citeas="ISO 7301:2011">
-                                  <locality type="clause">
-                                     <referenceFrom>3.1</referenceFrom>
-                                  </locality>
-                                  ISO\\u00a07301:2011, Clause 3.1
-                               </fmt-origin>
-                            </semx>
-                            , modified —
-                            <semx element="modification" source="_">The term "cargo rice" is shown as deprecated, and Note 1 to entry is not included here</semx>
-                         </semx>
-                         ]
-                      </p>
-                   </fmt-preferred>
-                   <domain id="_">rice</domain>
-                   <definition id="_">
-                      <verbal-definition>
-                         <p original-id="_">rice retaining its husk after threshing</p>
-                      </verbal-definition>
-                   </definition>
-                   <fmt-definition id="_">
-                      <semx element="definition" source="_">
-                         <p id="_">
-                            &lt;
-                            <semx element="domain" source="_">rice</semx>
-                            &gt; rice retaining its husk after threshing
-                         </p>
-                      </semx>
-                   </fmt-definition>
-                   <termexample id="_" keep-with-next="true" keep-lines-together="true" autonum="1">
-                      <fmt-name id="_">
-                         <span class="fmt-caption-label">
-                            <span class="fmt-element-name">EXAMPLE</span>
-                            <semx element="autonum" source="_">1</semx>
-                         </span>
-                      </fmt-name>
-                      <fmt-xref-label>
-                         <span class="fmt-element-name">Example</span>
-                         <semx element="autonum" source="_">1</semx>
-                      </fmt-xref-label>
-                      <fmt-xref-label container="paddy1">
-                         <span class="fmt-xref-container">
-                            <span class="fmt-element-name">Clause</span>
-                            <semx element="autonum" source="terms_and_definitions">1</semx>
-                            <span class="fmt-autonum-delim">.</span>
-                            <semx element="autonum" source="paddy1">1</semx>
-                         </span>
-                         <span class="fmt-comma">,</span>
-                         <span class="fmt-element-name">Example</span>
-                         <semx element="autonum" source="_">1</semx>
-                      </fmt-xref-label>
-                      <p id="_">Foreign seeds, husks, bran, sand, dust.</p>
-                      <ul>
-                         <li id="_">
-                            <fmt-name id="_">
-                               <semx element="autonum" source="_">—</semx>
-                            </fmt-name>
-                            A
-                         </li>
-                      </ul>
-                   </termexample>
-                   <termexample id="_" autonum="2">
-                      <fmt-name id="_">
-                         <span class="fmt-caption-label">
-                            <span class="fmt-element-name">EXAMPLE</span>
-                            <semx element="autonum" source="_">2</semx>
-                         </span>
-                      </fmt-name>
-                      <fmt-xref-label>
-                         <span class="fmt-element-name">Example</span>
-                         <semx element="autonum" source="_">2</semx>
-                      </fmt-xref-label>
-                      <fmt-xref-label container="paddy1">
-                         <span class="fmt-xref-container">
-                            <span class="fmt-element-name">Clause</span>
-                            <semx element="autonum" source="terms_and_definitions">1</semx>
-                            <span class="fmt-autonum-delim">.</span>
-                            <semx element="autonum" source="paddy1">1</semx>
-                         </span>
-                         <span class="fmt-comma">,</span>
-                         <span class="fmt-element-name">Example</span>
-                         <semx element="autonum" source="_">2</semx>
-                      </fmt-xref-label>
-                      <ul>
-                         <li id="_">
-                            <fmt-name id="_">
-                               <semx element="autonum" source="_">—</semx>
-                            </fmt-name>
-                            A
-                         </li>
-                      </ul>
-                   </termexample>
-                   <source status="identical" id="_">
-                      <origin citeas="">
-                         <termref base="IEV" target="xyz">t1</termref>
-                      </origin>
-                      <modification id="_">
-                         <p id="_">comment</p>
-                      </modification>
-                   </source>
-                   <source status="modified" id="_">
-                      <origin citeas="">
-                         <termref base="IEV" target="xyz"/>
-                      </origin>
-                      <modification id="_">
-                         <p original-id="_">with adjustments</p>
-                      </modification>
-                   </source>
-                   <source status="adapted" id="_">
-                      <origin citeas="">
-                         <termref base="IEV" target="xyz">t1</termref>
-                      </origin>
-                      <modification id="_">
-                         <p original-id="_"/>
-                      </modification>
-                   </source>
-                   <fmt-termsource status="identical">
-                      [SOURCE:
-                      <semx element="source" source="_">
-                         <origin citeas="" id="_">
-                            <termref base="IEV" target="xyz">t1</termref>
-                         </origin>
-                         <semx element="origin" source="_">
-                            <fmt-origin citeas="">
-                               <termref base="IEV" target="xyz">t1</termref>
-                            </fmt-origin>
-                         </semx>
-                         —
-                         <semx element="modification" source="_">comment</semx>
-                      </semx>
-                      ;
-                      <semx element="source" source="_">
-                         <origin citeas="" id="_">
-                            <termref base="IEV" target="xyz"/>
-                         </origin>
-                         <semx element="origin" source="_">
-                            <fmt-origin citeas="">
-                               <termref base="IEV" target="xyz"/>
-                            </fmt-origin>
-                         </semx>
-                         , modified —
-                         <semx element="modification" source="_">with adjustments</semx>
-                      </semx>
-                      ;
-                      <semx element="source" source="_">
-                         <origin citeas="" id="_">
-                            <termref base="IEV" target="xyz">t1</termref>
-                         </origin>
-                         <semx element="origin" source="_">
-                            <fmt-origin citeas="">
-                               <termref base="IEV" target="xyz">t1</termref>
-                            </fmt-origin>
-                         </semx>
-                         , adapted
-                      </semx>
-                      ]
-                   </fmt-termsource>
-                </term>
-                <term id="paddy">
-                   <fmt-name id="_">
-                      <span class="fmt-caption-label">
-                         <semx element="autonum" source="terms_and_definitions">1</semx>
-                         <span class="fmt-autonum-delim">.</span>
-                         <semx element="autonum" source="paddy">2</semx>
-                         <span class="fmt-autonum-delim">.</span>
-                      </span>
-                   </fmt-name>
-                   <fmt-xref-label>
-                      <span class="fmt-element-name">Clause</span>
-                      <semx element="autonum" source="terms_and_definitions">1</semx>
-                      <span class="fmt-autonum-delim">.</span>
-                      <semx element="autonum" source="paddy">2</semx>
-                   </fmt-xref-label>
-                   <preferred id="_">
-                      <expression>
-                         <name>paddy</name>
-                      </expression>
-                   </preferred>
-                   <fmt-preferred>
-                      <p>
-                         <semx element="preferred" source="_">
-                            <strong>paddy</strong>
-                         </semx>
-                      </p>
-                   </fmt-preferred>
-                   <admitted id="_">
-                      <letter-symbol>
-                         <name>paddy rice</name>
-                      </letter-symbol>
-                      <field-of-application id="_">in agriculture</field-of-application>
-                   </admitted>
-                   <admitted id="_">
-                      <expression>
-                         <name>rough rice</name>
-                      </expression>
-                   </admitted>
-                   <fmt-admitted>
-                      <p>
-                         <semx element="admitted" source="_">
-                            paddy rice
-                            <span class="fmt-designation-field">
-                               , &lt;
-                               <semx element="field-of-application" source="_">in agriculture</semx>
-                               &gt;
-                            </span>
-                         </semx>
-                      </p>
-                      <p>
-                         <semx element="admitted" source="_">rough rice</semx>
-                      </p>
-                   </fmt-admitted>
-                   <deprecates id="_">
-                      <expression>
-                         <name>cargo rice</name>
-                      </expression>
-                   </deprecates>
-                   <fmt-deprecates>
-                      <p>
-                         DEPRECATED:
-                         <semx element="deprecates" source="_">cargo rice</semx>
-                      </p>
-                   </fmt-deprecates>
-                   <definition id="_">
-                      <verbal-definition>
-                         <p original-id="_">rice retaining its husk after threshing</p>
-                      </verbal-definition>
-                   </definition>
-                   <fmt-definition id="_">
-                      <semx element="definition" source="_">
-                         <p id="_">rice retaining its husk after threshing</p>
-                      </semx>
-                   </fmt-definition>
-                   <termexample id="_" autonum="">
-                      <fmt-name id="_">
-                         <span class="fmt-caption-label">
-                            <span class="fmt-element-name">EXAMPLE</span>
-                         </span>
-                      </fmt-name>
-                      <fmt-xref-label>
-                         <span class="fmt-element-name">Example</span>
-                      </fmt-xref-label>
-                      <fmt-xref-label container="paddy">
-                         <span class="fmt-xref-container">
-                            <span class="fmt-element-name">Clause</span>
-                            <semx element="autonum" source="terms_and_definitions">1</semx>
-                            <span class="fmt-autonum-delim">.</span>
-                            <semx element="autonum" source="paddy">2</semx>
-                         </span>
-                         <span class="fmt-comma">,</span>
-                         <span class="fmt-element-name">Example</span>
-                      </fmt-xref-label>
-                      <ul>
-                         <li id="_">
-                            <fmt-name id="_">
-                               <semx element="autonum" source="_">—</semx>
-                            </fmt-name>
-                            A
-                         </li>
-                      </ul>
-                   </termexample>
-                   <termnote id="_" keep-with-next="true" keep-lines-together="true" autonum="1">
-                      <fmt-name id="_">
-                         <span class="fmt-caption-label">
-                            Note
-                            <semx element="autonum" source="_">1</semx>
-                            to entry
-                         </span>
-                         <span class="fmt-label-delim">: </span>
-                      </fmt-name>
-                      <fmt-xref-label>
-                         <span class="fmt-element-name">Note</span>
-                         <semx element="autonum" source="_">1</semx>
-                      </fmt-xref-label>
-                      <fmt-xref-label container="paddy">
-                         <span class="fmt-xref-container">
-                            <span class="fmt-element-name">Clause</span>
-                            <semx element="autonum" source="terms_and_definitions">1</semx>
-                            <span class="fmt-autonum-delim">.</span>
-                            <semx element="autonum" source="paddy">2</semx>
-                         </span>
-                         <span class="fmt-comma">,</span>
-                         <span class="fmt-element-name">Note</span>
-                         <semx element="autonum" source="_">1</semx>
-                      </fmt-xref-label>
-                      <p id="_">The starch of waxy rice consists almost entirely of amylopectin. The kernels have a tendency to stick together after cooking.</p>
-                   </termnote>
-                   <termnote id="_" autonum="2">
-                      <fmt-name id="_">
-                         <span class="fmt-caption-label">
-                            Note
-                            <semx element="autonum" source="_">2</semx>
-                            to entry
-                         </span>
-                         <span class="fmt-label-delim">: </span>
-                      </fmt-name>
-                      <fmt-xref-label>
-                         <span class="fmt-element-name">Note</span>
-                         <semx element="autonum" source="_">2</semx>
-                      </fmt-xref-label>
-                      <fmt-xref-label container="paddy">
-                         <span class="fmt-xref-container">
-                            <span class="fmt-element-name">Clause</span>
-                            <semx element="autonum" source="terms_and_definitions">1</semx>
-                            <span class="fmt-autonum-delim">.</span>
-                            <semx element="autonum" source="paddy">2</semx>
-                         </span>
-                         <span class="fmt-comma">,</span>
-                         <span class="fmt-element-name">Note</span>
-                         <semx element="autonum" source="_">2</semx>
-                      </fmt-xref-label>
-                      <ul>
-                         <li id="_">
-                            <fmt-name id="_">
-                               <semx element="autonum" source="_">—</semx>
-                            </fmt-name>
-                            A
-                         </li>
-                      </ul>
-                      <p id="_">The starch of waxy rice consists almost entirely of amylopectin. The kernels have a tendency to stick together after cooking.</p>
-                   </termnote>
-                   <source status="identical" id="_">
-                      <origin bibitemid="ISO7301" type="inline" droploc="true" citeas="ISO 7301:2011">
-                         <locality type="clause">
-                            <referenceFrom>3.1</referenceFrom>
-                         </locality>
-                      </origin>
-                      <origin bibitemid="ISO7301" type="inline" case="lowercase" citeas="ISO 7301:2011">
-                         <locality type="clause">
-                            <referenceFrom>3.1</referenceFrom>
-                         </locality>
-                      </origin>
-                   </source>
-                   <fmt-termsource status="identical">
-                      [SOURCE:
-                      <semx element="source" source="_">
-                         <origin bibitemid="ISO7301" type="inline" droploc="true" citeas="ISO 7301:2011" id="_">
-                            <locality type="clause">
-                               <referenceFrom>3.1</referenceFrom>
-                            </locality>
-                         </origin>
-                         <semx element="origin" source="_">
-                            <fmt-origin bibitemid="ISO7301" type="inline" droploc="true" citeas="ISO 7301:2011">
-                               <locality type="clause">
-                                  <referenceFrom>3.1</referenceFrom>
-                               </locality>
-                               ISO\\u00a07301:2011, 3.1
-                            </fmt-origin>
-                         </semx>
-                         <origin bibitemid="ISO7301" type="inline" case="lowercase" citeas="ISO 7301:2011" id="_">
-                            <locality type="clause">
-                               <referenceFrom>3.1</referenceFrom>
-                            </locality>
-                         </origin>
-                         <semx element="origin" source="_">
-                            <fmt-origin bibitemid="ISO7301" type="inline" case="lowercase" citeas="ISO 7301:2011">
-                               <locality type="clause">
-                                  <referenceFrom>3.1</referenceFrom>
-                               </locality>
-                               ISO\\u00a07301:2011, clause 3.1
-                            </fmt-origin>
-                         </semx>
-                      </semx>
-                      ]
-                   </fmt-termsource>
-                </term>
-             </terms>
-          </sections>
-       </iso-standard>
+      <iso-standard xmlns="http://riboseinc.com/isoxml" type="presentation">
+         <preface>
+            <clause type="toc" id="_" displayorder="1">
+               <fmt-title id="_" depth="1">Table of contents</fmt-title>
+            </clause>
+         </preface>
+         <sections>
+            <terms id="terms_and_definitions" obligation="normative" displayorder="2">
+               <title id="_">Terms and Definitions</title>
+               <fmt-title id="_" depth="1">
+                  <span class="fmt-caption-label">
+                     <semx element="autonum" source="terms_and_definitions">1</semx>
+                     <span class="fmt-autonum-delim">.</span>
+                  </span>
+                  <span class="fmt-caption-delim">
+                     <tab/>
+                  </span>
+                  <semx element="title" source="_">Terms and Definitions</semx>
+               </fmt-title>
+               <fmt-xref-label>
+                  <span class="fmt-element-name">Clause</span>
+                  <semx element="autonum" source="terms_and_definitions">1</semx>
+               </fmt-xref-label>
+               <p>For the purposes of this document, the following terms and definitions apply.</p>
+               <term id="paddy1">
+                  <fmt-name id="_">
+                     <span class="fmt-caption-label">
+                        <semx element="autonum" source="terms_and_definitions">1</semx>
+                        <span class="fmt-autonum-delim">.</span>
+                        <semx element="autonum" source="paddy1">1</semx>
+                        <span class="fmt-autonum-delim">.</span>
+                     </span>
+                  </fmt-name>
+                  <fmt-xref-label>
+                     <span class="fmt-element-name">Clause</span>
+                     <semx element="autonum" source="terms_and_definitions">1</semx>
+                     <span class="fmt-autonum-delim">.</span>
+                     <semx element="autonum" source="paddy1">1</semx>
+                  </fmt-xref-label>
+                  <preferred id="_">
+                     <expression>
+                        <name>paddy</name>
+                     </expression>
+                     <field-of-application id="_">in agriculture</field-of-application>
+                     <usage-info id="_">dated</usage-info>
+                     <source status="modified" id="_">
+                        <origin bibitemid="ISO7301" type="inline" citeas="ISO 7301:2011">
+                           <locality type="clause">
+                              <referenceFrom>3.1</referenceFrom>
+                           </locality>
+                        </origin>
+                        <modification id="_">
+                           <p id="_">The term "cargo rice" is shown as deprecated, and Note 1 to entry is not included here</p>
+                        </modification>
+                     </source>
+                  </preferred>
+                  <fmt-preferred>
+                     <p>
+                        <semx element="preferred" source="_">
+                           <strong>paddy</strong>
+                           <span class="fmt-designation-field">
+                              , &lt;
+                              <semx element="field-of-application" source="_">in agriculture</semx>
+                              ,
+                              <semx element="usage-info" source="_">dated</semx>
+                              &gt;
+                           </span>
+                        </semx>
+                        [SOURCE:
+                        <semx element="source" source="_">
+                           <origin bibitemid="ISO7301" type="inline" citeas="ISO 7301:2011" id="_">
+                              <locality type="clause">
+                                 <referenceFrom>3.1</referenceFrom>
+                              </locality>
+                           </origin>
+                           <semx element="origin" source="_">
+                              <fmt-origin bibitemid="ISO7301" type="inline" citeas="ISO 7301:2011">
+                                 <locality type="clause">
+                                    <referenceFrom>3.1</referenceFrom>
+                                 </locality>
+                                 ISO\\u00a07301:2011, Clause 3.1
+                              </fmt-origin>
+                           </semx>
+                           , modified —
+                           <semx element="modification" source="_">The term "cargo rice" is shown as deprecated, and Note 1 to entry is not included here</semx>
+                        </semx>
+                        ]
+                     </p>
+                  </fmt-preferred>
+                  <domain id="_">rice</domain>
+                  <definition id="_">
+                     <verbal-definition>
+                        <p original-id="_">rice retaining its husk after threshing</p>
+                     </verbal-definition>
+                  </definition>
+                  <fmt-definition id="_">
+                     <semx element="definition" source="_">
+                        <p id="_">
+                           &lt;
+                           <semx element="domain" source="_">rice</semx>
+                           &gt; rice retaining its husk after threshing
+                        </p>
+                     </semx>
+                  </fmt-definition>
+                  <termexample id="_" keep-with-next="true" keep-lines-together="true" autonum="1">
+                     <fmt-name id="_">
+                        <span class="fmt-caption-label">
+                           <span class="fmt-element-name">EXAMPLE</span>
+                           <semx element="autonum" source="_">1</semx>
+                        </span>
+                     </fmt-name>
+                     <fmt-xref-label>
+                        <span class="fmt-element-name">Example</span>
+                        <semx element="autonum" source="_">1</semx>
+                     </fmt-xref-label>
+                     <fmt-xref-label container="paddy1">
+                        <span class="fmt-xref-container">
+                           <span class="fmt-element-name">Clause</span>
+                           <semx element="autonum" source="terms_and_definitions">1</semx>
+                           <span class="fmt-autonum-delim">.</span>
+                           <semx element="autonum" source="paddy1">1</semx>
+                        </span>
+                        <span class="fmt-comma">,</span>
+                        <span class="fmt-element-name">Example</span>
+                        <semx element="autonum" source="_">1</semx>
+                     </fmt-xref-label>
+                     <p id="_">Foreign seeds, husks, bran, sand, dust.</p>
+                     <ul>
+                        <li id="_">
+                           <fmt-name id="_">
+                              <semx element="autonum" source="_">—</semx>
+                           </fmt-name>
+                           A
+                        </li>
+                     </ul>
+                  </termexample>
+                  <termexample id="_" autonum="2">
+                     <fmt-name id="_">
+                        <span class="fmt-caption-label">
+                           <span class="fmt-element-name">EXAMPLE</span>
+                           <semx element="autonum" source="_">2</semx>
+                        </span>
+                     </fmt-name>
+                     <fmt-xref-label>
+                        <span class="fmt-element-name">Example</span>
+                        <semx element="autonum" source="_">2</semx>
+                     </fmt-xref-label>
+                     <fmt-xref-label container="paddy1">
+                        <span class="fmt-xref-container">
+                           <span class="fmt-element-name">Clause</span>
+                           <semx element="autonum" source="terms_and_definitions">1</semx>
+                           <span class="fmt-autonum-delim">.</span>
+                           <semx element="autonum" source="paddy1">1</semx>
+                        </span>
+                        <span class="fmt-comma">,</span>
+                        <span class="fmt-element-name">Example</span>
+                        <semx element="autonum" source="_">2</semx>
+                     </fmt-xref-label>
+                     <ul>
+                        <li id="_">
+                           <fmt-name id="_">
+                              <semx element="autonum" source="_">—</semx>
+                           </fmt-name>
+                           A
+                        </li>
+                     </ul>
+                  </termexample>
+                  <source status="identical" id="_">
+                     <origin citeas="">
+                        <termref base="IEV" target="xyz">t1</termref>
+                     </origin>
+                     <modification id="_">
+                        <p id="_">comment</p>
+                     </modification>
+                  </source>
+                  <source status="modified" id="_">
+                     <origin citeas="">
+                        <termref base="IEV" target="xyz"/>
+                     </origin>
+                     <modification id="_">
+                        <p original-id="_">with adjustments</p>
+                     </modification>
+                  </source>
+                  <source status="adapted" id="_">
+                     <origin citeas="">
+                        <termref base="IEV" target="xyz">t1</termref>
+                     </origin>
+                     <modification id="_">
+                        <p original-id="_"/>
+                     </modification>
+                  </source>
+                  <fmt-termsource status="identical">
+                     [SOURCE:
+                     <semx element="source" source="_">
+                        <origin citeas="" id="_">
+                           <termref base="IEV" target="xyz">t1</termref>
+                        </origin>
+                        <semx element="origin" source="_">
+                           <fmt-origin citeas="">
+                              <termref base="IEV" target="xyz">t1</termref>
+                           </fmt-origin>
+                        </semx>
+                        —
+                        <semx element="modification" source="_">comment</semx>
+                     </semx>
+                     ;
+                     <semx element="source" source="_">
+                        <origin citeas="" id="_">
+                           <termref base="IEV" target="xyz"/>
+                        </origin>
+                        <semx element="origin" source="_">
+                           <fmt-origin citeas="">
+                              <termref base="IEV" target="xyz"/>
+                           </fmt-origin>
+                        </semx>
+                        , modified —
+                        <semx element="modification" source="_">with adjustments</semx>
+                     </semx>
+                     ;
+                     <semx element="source" source="_">
+                        <origin citeas="" id="_">
+                           <termref base="IEV" target="xyz">t1</termref>
+                        </origin>
+                        <semx element="origin" source="_">
+                           <fmt-origin citeas="">
+                              <termref base="IEV" target="xyz">t1</termref>
+                           </fmt-origin>
+                        </semx>
+                        , adapted
+                     </semx>
+                     ]
+                  </fmt-termsource>
+               </term>
+               <term id="paddy">
+                  <fmt-name id="_">
+                     <span class="fmt-caption-label">
+                        <semx element="autonum" source="terms_and_definitions">1</semx>
+                        <span class="fmt-autonum-delim">.</span>
+                        <semx element="autonum" source="paddy">2</semx>
+                        <span class="fmt-autonum-delim">.</span>
+                     </span>
+                  </fmt-name>
+                  <fmt-xref-label>
+                     <span class="fmt-element-name">Clause</span>
+                     <semx element="autonum" source="terms_and_definitions">1</semx>
+                     <span class="fmt-autonum-delim">.</span>
+                     <semx element="autonum" source="paddy">2</semx>
+                  </fmt-xref-label>
+                  <preferred id="_">
+                     <expression>
+                        <name>paddy</name>
+                     </expression>
+                  </preferred>
+                  <fmt-preferred>
+                     <p>
+                        <semx element="preferred" source="_">
+                           <strong>paddy</strong>
+                        </semx>
+                     </p>
+                  </fmt-preferred>
+                  <admitted id="_">
+                     <letter-symbol>
+                        <name>paddy rice</name>
+                     </letter-symbol>
+                     <field-of-application id="_">in agriculture</field-of-application>
+                  </admitted>
+                  <admitted id="_">
+                     <expression>
+                        <name>rough rice</name>
+                     </expression>
+                  </admitted>
+                  <fmt-admitted>
+                     <p>
+                        <semx element="admitted" source="_">
+                           paddy rice
+                           <span class="fmt-designation-field">
+                              , &lt;
+                              <semx element="field-of-application" source="_">in agriculture</semx>
+                              &gt;
+                           </span>
+                        </semx>
+                     </p>
+                     <p>
+                        <semx element="admitted" source="_">rough rice</semx>
+                     </p>
+                  </fmt-admitted>
+                  <deprecates id="_">
+                     <expression>
+                        <name>cargo rice</name>
+                     </expression>
+                  </deprecates>
+                  <fmt-deprecates>
+                     <p>
+                        DEPRECATED:
+                        <semx element="deprecates" source="_">cargo rice</semx>
+                     </p>
+                  </fmt-deprecates>
+                  <definition id="_">
+                     <verbal-definition>
+                        <p original-id="_">rice retaining its husk after threshing</p>
+                     </verbal-definition>
+                  </definition>
+                  <fmt-definition id="_">
+                     <semx element="definition" source="_">
+                        <p id="_">rice retaining its husk after threshing</p>
+                     </semx>
+                  </fmt-definition>
+                  <termexample id="_" autonum="">
+                     <fmt-name id="_">
+                        <span class="fmt-caption-label">
+                           <span class="fmt-element-name">EXAMPLE</span>
+                        </span>
+                     </fmt-name>
+                     <fmt-xref-label>
+                        <span class="fmt-element-name">Example</span>
+                     </fmt-xref-label>
+                     <fmt-xref-label container="paddy">
+                        <span class="fmt-xref-container">
+                           <span class="fmt-element-name">Clause</span>
+                           <semx element="autonum" source="terms_and_definitions">1</semx>
+                           <span class="fmt-autonum-delim">.</span>
+                           <semx element="autonum" source="paddy">2</semx>
+                        </span>
+                        <span class="fmt-comma">,</span>
+                        <span class="fmt-element-name">Example</span>
+                     </fmt-xref-label>
+                     <ul>
+                        <li id="_">
+                           <fmt-name id="_">
+                              <semx element="autonum" source="_">—</semx>
+                           </fmt-name>
+                           A
+                        </li>
+                     </ul>
+                  </termexample>
+                  <termnote id="_" keep-with-next="true" keep-lines-together="true" autonum="1">
+                     <fmt-name id="_">
+                        <span class="fmt-caption-label">
+                           Note
+                           <semx element="autonum" source="_">1</semx>
+                           to entry
+                        </span>
+                        <span class="fmt-label-delim">: </span>
+                     </fmt-name>
+                     <fmt-xref-label>
+                        <span class="fmt-element-name">Note</span>
+                        <semx element="autonum" source="_">1</semx>
+                     </fmt-xref-label>
+                     <fmt-xref-label container="paddy">
+                        <span class="fmt-xref-container">
+                           <span class="fmt-element-name">Clause</span>
+                           <semx element="autonum" source="terms_and_definitions">1</semx>
+                           <span class="fmt-autonum-delim">.</span>
+                           <semx element="autonum" source="paddy">2</semx>
+                        </span>
+                        <span class="fmt-comma">,</span>
+                        <span class="fmt-element-name">Note</span>
+                        <semx element="autonum" source="_">1</semx>
+                     </fmt-xref-label>
+                     <p id="_">The starch of waxy rice consists almost entirely of amylopectin. The kernels have a tendency to stick together after cooking.</p>
+                  </termnote>
+                  <termnote id="_" autonum="2">
+                     <fmt-name id="_">
+                        <span class="fmt-caption-label">
+                           Note
+                           <semx element="autonum" source="_">2</semx>
+                           to entry
+                        </span>
+                        <span class="fmt-label-delim">: </span>
+                     </fmt-name>
+                     <fmt-xref-label>
+                        <span class="fmt-element-name">Note</span>
+                        <semx element="autonum" source="_">2</semx>
+                     </fmt-xref-label>
+                     <fmt-xref-label container="paddy">
+                        <span class="fmt-xref-container">
+                           <span class="fmt-element-name">Clause</span>
+                           <semx element="autonum" source="terms_and_definitions">1</semx>
+                           <span class="fmt-autonum-delim">.</span>
+                           <semx element="autonum" source="paddy">2</semx>
+                        </span>
+                        <span class="fmt-comma">,</span>
+                        <span class="fmt-element-name">Note</span>
+                        <semx element="autonum" source="_">2</semx>
+                     </fmt-xref-label>
+                     <ul>
+                        <li id="_">
+                           <fmt-name id="_">
+                              <semx element="autonum" source="_">—</semx>
+                           </fmt-name>
+                           A
+                        </li>
+                     </ul>
+                     <p id="_">The starch of waxy rice consists almost entirely of amylopectin. The kernels have a tendency to stick together after cooking.</p>
+                  </termnote>
+                  <source status="identical" id="_">
+                     <origin bibitemid="ISO7301" type="inline" droploc="true" citeas="ISO 7301:2011">
+                        <locality type="clause">
+                           <referenceFrom>3.1</referenceFrom>
+                        </locality>
+                     </origin>
+                     <origin bibitemid="ISO7301" type="inline" case="lowercase" citeas="ISO 7301:2011">
+                        <locality type="clause">
+                           <referenceFrom>3.1</referenceFrom>
+                        </locality>
+                     </origin>
+                  </source>
+                  <fmt-termsource status="identical">
+                     [SOURCE:
+                     <semx element="source" source="_">
+                        <origin bibitemid="ISO7301" type="inline" droploc="true" citeas="ISO 7301:2011" id="_">
+                           <locality type="clause">
+                              <referenceFrom>3.1</referenceFrom>
+                           </locality>
+                        </origin>
+                        <semx element="origin" source="_">
+                           <fmt-origin bibitemid="ISO7301" type="inline" droploc="true" citeas="ISO 7301:2011">
+                              <locality type="clause">
+                                 <referenceFrom>3.1</referenceFrom>
+                              </locality>
+                              ISO\\u00a07301:2011, 3.1
+                           </fmt-origin>
+                        </semx>
+                        <origin bibitemid="ISO7301" type="inline" case="lowercase" citeas="ISO 7301:2011" id="_">
+                           <locality type="clause">
+                              <referenceFrom>3.1</referenceFrom>
+                           </locality>
+                        </origin>
+                        <semx element="origin" source="_">
+                           <fmt-origin bibitemid="ISO7301" type="inline" case="lowercase" citeas="ISO 7301:2011">
+                              <locality type="clause">
+                                 <referenceFrom>3.1</referenceFrom>
+                              </locality>
+                              ISO\\u00a07301:2011, clause 3.1
+                           </fmt-origin>
+                        </semx>
+                     </semx>
+                     ]
+                  </fmt-termsource>
+               </term>
+            </terms>
+         </sections>
+      </iso-standard>
     PRESXML
 
     html = <<~"OUTPUT"
@@ -1049,584 +1049,6 @@ RSpec.describe IsoDoc do
       .to be_equivalent_to Canon.format_xml(presxml)
   end
 
-  it "processes IsoXML term with multiple preferred terms" do
-    input = <<~INPUT
-          <iso-standard xmlns="http://riboseinc.com/isoxml">
-          <sections>
-          <terms id="terms_and_definitions" obligation="normative"><title>Terms and Definitions</title>
-      <term id="paddy1">
-      <preferred><expression><name>paddy</name></expression></preferred>
-      <preferred><expression><name>muddy rice</name></expression></preferred>
-      <domain>rice</domain>
-      <definition><verbal-definition><p id="_eb29b35e-123e-4d1c-b50b-2714d41e747f">rice retaining its husk after threshing</p></verbal-definition></definition>
-      </term>
-      <term id="paddy2">
-      <preferred><expression><name language="eng">paddy</name></expression></preferred>
-      <preferred><expression><name language="eng">muddy rice</name></expression></preferred>
-      <domain>rice</domain>
-      <definition><verbal-definition><p id="_eb29b35e-123e-4d1c-b50b-2714d41e747a">rice retaining its husk after threshing</p></verbal-definition></definition>
-      </term>
-      <term id="paddy3">
-      <preferred geographic-area="US"><expression><name>paddy</name></expression></preferred>
-      <preferred><expression><name>muddy rice</name></expression></preferred>
-      <domain>rice</domain>
-      <definition><verbal-definition><p id="_eb29b35e-123e-4d1c-b50b-2714d41e747b">rice retaining its husk after threshing</p></verbal-definition></definition>
-      </term>
-            <term id="paddy4">
-      <preferred><expression language="eng"><name>paddy</name></expression></preferred>
-      <preferred><expression language="fra"><name>muddy rice</name></expression></preferred>
-      <domain>rice</domain>
-      <definition><verbal-definition><p id="_eb29b35e-123e-4d1c-b50b-2714d41e747c">rice retaining its husk after threshing</p></verbal-definition></definition>
-      </term>
-                </terms>
-        </sections>
-      </iso-standard>
-    INPUT
-    presxml = <<~PRESXML
-      <terms id="terms_and_definitions" obligation="normative" displayorder="2">
-         <title id="_">Terms and Definitions</title>
-         <fmt-title id="_" depth="1">
-            <span class="fmt-caption-label">
-               <semx element="autonum" source="terms_and_definitions">1</semx>
-               <span class="fmt-autonum-delim">.</span>
-            </span>
-            <span class="fmt-caption-delim">
-               <tab/>
-            </span>
-            <semx element="title" source="_">Terms and Definitions</semx>
-         </fmt-title>
-         <fmt-xref-label>
-            <span class="fmt-element-name">Clause</span>
-            <semx element="autonum" source="terms_and_definitions">1</semx>
-         </fmt-xref-label>
-         <term id="paddy1">
-            <fmt-name id="_">
-               <span class="fmt-caption-label">
-                  <semx element="autonum" source="terms_and_definitions">1</semx>
-                  <span class="fmt-autonum-delim">.</span>
-                  <semx element="autonum" source="paddy1">1</semx>
-                  <span class="fmt-autonum-delim">.</span>
-               </span>
-            </fmt-name>
-            <fmt-xref-label>
-               <span class="fmt-element-name">Clause</span>
-               <semx element="autonum" source="terms_and_definitions">1</semx>
-               <span class="fmt-autonum-delim">.</span>
-               <semx element="autonum" source="paddy1">1</semx>
-            </fmt-xref-label>
-            <preferred id="_">
-               <expression>
-                  <name>paddy</name>
-               </expression>
-            </preferred>
-            <preferred id="_">
-               <expression>
-                  <name>muddy rice</name>
-               </expression>
-            </preferred>
-            <fmt-preferred>
-               <p>
-                  <semx element="preferred" source="_">
-                     <strong>paddy</strong>
-                  </semx>
-                  ;
-                  <semx element="preferred" source="_">
-                     <strong>muddy rice</strong>
-                  </semx>
-               </p>
-            </fmt-preferred>
-            <domain id="_">rice</domain>
-            <definition id="_">
-               <verbal-definition>
-                  <p original-id="_">rice retaining its husk after threshing</p>
-               </verbal-definition>
-            </definition>
-            <fmt-definition id="_">
-               <semx element="definition" source="_">
-                  <p id="_">
-                     &lt;
-                     <semx element="domain" source="_">rice</semx>
-                     &gt; rice retaining its husk after threshing
-                  </p>
-               </semx>
-            </fmt-definition>
-         </term>
-         <term id="paddy2">
-            <fmt-name id="_">
-               <span class="fmt-caption-label">
-                  <semx element="autonum" source="terms_and_definitions">1</semx>
-                  <span class="fmt-autonum-delim">.</span>
-                  <semx element="autonum" source="paddy2">2</semx>
-                  <span class="fmt-autonum-delim">.</span>
-               </span>
-            </fmt-name>
-            <fmt-xref-label>
-               <span class="fmt-element-name">Clause</span>
-               <semx element="autonum" source="terms_and_definitions">1</semx>
-               <span class="fmt-autonum-delim">.</span>
-               <semx element="autonum" source="paddy2">2</semx>
-            </fmt-xref-label>
-            <preferred id="_">
-               <expression>
-                  <name language="eng">paddy</name>
-               </expression>
-            </preferred>
-            <preferred id="_">
-               <expression>
-                  <name language="eng">muddy rice</name>
-               </expression>
-            </preferred>
-            <fmt-preferred>
-               <p>
-                  <semx element="preferred" source="_">
-                     <strong>paddy</strong>
-                  </semx>
-                  ;
-                  <semx element="preferred" source="_">
-                     <strong>muddy rice</strong>
-                  </semx>
-               </p>
-            </fmt-preferred>
-            <domain id="_">rice</domain>
-            <definition id="_">
-               <verbal-definition>
-                  <p original-id="_">rice retaining its husk after threshing</p>
-               </verbal-definition>
-            </definition>
-            <fmt-definition id="_">
-               <semx element="definition" source="_">
-                  <p id="_">
-                     &lt;
-                     <semx element="domain" source="_">rice</semx>
-                     &gt; rice retaining its husk after threshing
-                  </p>
-               </semx>
-            </fmt-definition>
-         </term>
-         <term id="paddy3">
-            <fmt-name id="_">
-               <span class="fmt-caption-label">
-                  <semx element="autonum" source="terms_and_definitions">1</semx>
-                  <span class="fmt-autonum-delim">.</span>
-                  <semx element="autonum" source="paddy3">3</semx>
-                  <span class="fmt-autonum-delim">.</span>
-               </span>
-            </fmt-name>
-            <fmt-xref-label>
-               <span class="fmt-element-name">Clause</span>
-               <semx element="autonum" source="terms_and_definitions">1</semx>
-               <span class="fmt-autonum-delim">.</span>
-               <semx element="autonum" source="paddy3">3</semx>
-            </fmt-xref-label>
-            <preferred geographic-area="US" id="_">
-               <expression>
-                  <name>paddy</name>
-               </expression>
-            </preferred>
-            <preferred id="_">
-               <expression>
-                  <name>muddy rice</name>
-               </expression>
-            </preferred>
-            <fmt-preferred>
-               <p>
-                  <semx element="preferred" source="_">
-                     <strong>paddy</strong>
-                     , US
-                  </semx>
-               </p>
-               <p>
-                  <semx element="preferred" source="_">
-                     <strong>muddy rice</strong>
-                  </semx>
-               </p>
-            </fmt-preferred>
-            <domain id="_">rice</domain>
-            <definition id="_">
-               <verbal-definition>
-                  <p original-id="_">rice retaining its husk after threshing</p>
-               </verbal-definition>
-            </definition>
-            <fmt-definition id="_">
-               <semx element="definition" source="_">
-                  <p id="_">
-                     &lt;
-                     <semx element="domain" source="_">rice</semx>
-                     &gt; rice retaining its husk after threshing
-                  </p>
-               </semx>
-            </fmt-definition>
-         </term>
-         <term id="paddy4">
-            <fmt-name id="_">
-               <span class="fmt-caption-label">
-                  <semx element="autonum" source="terms_and_definitions">1</semx>
-                  <span class="fmt-autonum-delim">.</span>
-                  <semx element="autonum" source="paddy4">4</semx>
-                  <span class="fmt-autonum-delim">.</span>
-               </span>
-            </fmt-name>
-            <fmt-xref-label>
-               <span class="fmt-element-name">Clause</span>
-               <semx element="autonum" source="terms_and_definitions">1</semx>
-               <span class="fmt-autonum-delim">.</span>
-               <semx element="autonum" source="paddy4">4</semx>
-            </fmt-xref-label>
-            <preferred id="_">
-               <expression language="eng">
-                  <name>paddy</name>
-               </expression>
-            </preferred>
-            <preferred id="_">
-               <expression language="fra">
-                  <name>muddy rice</name>
-               </expression>
-            </preferred>
-            <fmt-preferred>
-               <p>
-                  <semx element="preferred" source="_">
-                     <strong>paddy</strong>
-                     , eng
-                  </semx>
-               </p>
-               <p>
-                  <semx element="preferred" source="_">
-                     <strong>muddy rice</strong>
-                     , fra
-                  </semx>
-               </p>
-            </fmt-preferred>
-            <domain id="_">rice</domain>
-            <definition id="_">
-               <verbal-definition>
-                  <p original-id="_">rice retaining its husk after threshing</p>
-               </verbal-definition>
-            </definition>
-            <fmt-definition id="_">
-               <semx element="definition" source="_">
-                  <p id="_">
-                     &lt;
-                     <semx element="domain" source="_">rice</semx>
-                     &gt; rice retaining its husk after threshing
-                  </p>
-               </semx>
-            </fmt-definition>
-         </term>
-      </terms>
-    PRESXML
-    expect(strip_guid(Canon.format_xml(Nokogiri::XML(IsoDoc::PresentationXMLConvert
-      .new(presxml_options)
-       .convert("test", input, true))
-      .at("//xmlns:terms").to_xml)))
-      .to be_equivalent_to Canon.format_xml(presxml)
-  end
-
-  it "processes IsoXML term with grammatical information" do
-    input = <<~INPUT
-          <iso-standard xmlns="http://riboseinc.com/isoxml">
-          <sections>
-          <terms id="terms_and_definitions" obligation="normative"><title>Terms and Definitions</title>
-      <term id="paddy1">
-      <preferred geographic-area="US"><expression language="en" script="Latn"><name>paddy</name>
-         <pronunciation>pædiː</pronunciation>
-                  <grammar>
-              <gender>masculine</gender>
-              <gender>feminine</gender>
-              <number>singular</number>
-              <isPreposition>false</isPreposition>
-              <isNoun>true</isNoun>
-              <grammar-value>irregular declension</grammar-value>
-            </grammar>
-      </expression></preferred>
-      <preferred><expression><name>muddy rice</name>
-                        <grammar>
-              <gender>neuter</gender>
-              <isNoun>true</isNoun>
-              <grammar-value>irregular declension</grammar-value>
-            </grammar>
-      </expression></preferred>
-      <domain>rice</domain>
-      <definition><verbal-definition><p id="_eb29b35e-123e-4d1c-b50b-2714d41e747f">rice retaining its husk after threshing</p></verbal-definition></definition>
-      </term>
-                </terms>
-        </sections>
-      </iso-standard>
-    INPUT
-    presxml = <<~PRESXML
-      <terms id="terms_and_definitions" obligation="normative" displayorder="2">
-         <title id="_">Terms and Definitions</title>
-         <fmt-title id="_" depth="1">
-            <span class="fmt-caption-label">
-               <semx element="autonum" source="terms_and_definitions">1</semx>
-               <span class="fmt-autonum-delim">.</span>
-            </span>
-            <span class="fmt-caption-delim">
-               <tab/>
-            </span>
-            <semx element="title" source="_">Terms and Definitions</semx>
-         </fmt-title>
-         <fmt-xref-label>
-            <span class="fmt-element-name">Clause</span>
-            <semx element="autonum" source="terms_and_definitions">1</semx>
-         </fmt-xref-label>
-         <term id="paddy1">
-            <fmt-name id="_">
-               <span class="fmt-caption-label">
-                  <semx element="autonum" source="terms_and_definitions">1</semx>
-                  <span class="fmt-autonum-delim">.</span>
-                  <semx element="autonum" source="paddy1">1</semx>
-                  <span class="fmt-autonum-delim">.</span>
-               </span>
-            </fmt-name>
-            <fmt-xref-label>
-               <span class="fmt-element-name">Clause</span>
-               <semx element="autonum" source="terms_and_definitions">1</semx>
-               <span class="fmt-autonum-delim">.</span>
-               <semx element="autonum" source="paddy1">1</semx>
-            </fmt-xref-label>
-            <preferred geographic-area="US" id="_">
-               <expression language="en" script="Latn">
-                  <name>paddy</name>
-                  <pronunciation>pædiː</pronunciation>
-                  <grammar>
-                     <gender>masculine</gender>
-                     <gender>feminine</gender>
-                     <number>singular</number>
-                     <isPreposition>false</isPreposition>
-                     <isNoun>true</isNoun>
-                     <grammar-value>irregular declension</grammar-value>
-                  </grammar>
-               </expression>
-            </preferred>
-            <preferred id="_">
-               <expression>
-                  <name>muddy rice</name>
-                  <grammar>
-                     <gender>neuter</gender>
-                     <isNoun>true</isNoun>
-                     <grammar-value>irregular declension</grammar-value>
-                  </grammar>
-               </expression>
-            </preferred>
-            <fmt-preferred>
-            <p>
-               <semx element="preferred" source="_">
-                  <strong>paddy</strong>
-                  , m, f, sg, noun, en Latn US, /pædiː/
-               </semx>
-            </p>
-            <p>
-               <semx element="preferred" source="_">
-                  <strong>muddy rice</strong>
-                  , n, noun
-               </semx>
-            </p>
-            </fmt-preferred>
-            <domain id="_">rice</domain>
-            <definition id="_">
-               <verbal-definition>
-                  <p original-id="_">rice retaining its husk after threshing</p>
-               </verbal-definition>
-            </definition>
-            <fmt-definition id="_">
-               <semx element="definition" source="_">
-                  <p id="_">
-                     &lt;
-                     <semx element="domain" source="_">rice</semx>
-                     &gt; rice retaining its husk after threshing
-                  </p>
-               </semx>
-            </fmt-definition>
-         </term>
-      </terms>
-    PRESXML
-    expect(strip_guid(Canon.format_xml(Nokogiri::XML(IsoDoc::PresentationXMLConvert
-      .new(presxml_options)
-       .convert("test", input, true))
-      .at("//xmlns:terms").to_xml)))
-      .to be_equivalent_to Canon.format_xml(presxml)
-  end
-
-  it "processes IsoXML term with empty, mathematical, or graphical designations" do
-    input = <<~INPUT
-          <iso-standard xmlns="http://riboseinc.com/isoxml">
-          <sections>
-          <terms id="terms_and_definitions" obligation="normative"><title>Terms and Definitions</title>
-      <term id="paddy1">
-      <preferred><expression><name/></expression></preferred>
-      <preferred isInternational='true'><graphical-symbol><figure id='_'><pre id='_'>&lt;LITERAL&gt; FIGURATIVE</pre></figure>
-                 </graphical-symbol>
-               </preferred>
-               <preferred>
-                 <letter-symbol>
-                   <name>
-                     <stem type='MathML'>
-                       <math xmlns='http://www.w3.org/1998/Math/MathML'>
-                         <msub>
-                           <mrow>
-                             <mi>x</mi>
-                           </mrow>
-                           <mrow>
-                             <mn>1</mn>
-                           </mrow>
-                         </msub>
-                       </math>
-                     </stem>
-                   </name>
-                 </letter-symbol>
-               </preferred>
-      <domain>rice</domain>
-      <definition><verbal-definition><p id="_eb29b35e-123e-4d1c-b50b-2714d41e747f">rice retaining its husk after threshing</p></verbal-definition></definition>
-      </term>
-                </terms>
-        </sections>
-      </iso-standard>
-    INPUT
-    presxml = <<~PRESXML
-      <terms id="terms_and_definitions" obligation="normative" displayorder="2">
-         <title id="_">Terms and Definitions</title>
-         <fmt-title id="_" depth="1">
-            <span class="fmt-caption-label">
-               <semx element="autonum" source="terms_and_definitions">1</semx>
-               <span class="fmt-autonum-delim">.</span>
-            </span>
-            <span class="fmt-caption-delim">
-               <tab/>
-            </span>
-            <semx element="title" source="_">Terms and Definitions</semx>
-         </fmt-title>
-         <fmt-xref-label>
-            <span class="fmt-element-name">Clause</span>
-            <semx element="autonum" source="terms_and_definitions">1</semx>
-         </fmt-xref-label>
-         <term id="paddy1">
-            <fmt-name id="_">
-               <span class="fmt-caption-label">
-                  <semx element="autonum" source="terms_and_definitions">1</semx>
-                  <span class="fmt-autonum-delim">.</span>
-                  <semx element="autonum" source="paddy1">1</semx>
-                  <span class="fmt-autonum-delim">.</span>
-               </span>
-            </fmt-name>
-            <fmt-xref-label>
-               <span class="fmt-element-name">Clause</span>
-               <semx element="autonum" source="terms_and_definitions">1</semx>
-               <span class="fmt-autonum-delim">.</span>
-               <semx element="autonum" source="paddy1">1</semx>
-            </fmt-xref-label>
-            <preferred id="_">
-               <expression>
-                  <name/>
-               </expression>
-            </preferred>
-            <preferred isInternational="true" id="_">
-               <graphical-symbol>
-                  <figure autonum="1" original-id="_">
-                     <pre original-id="_">&lt;LITERAL&gt; FIGURATIVE</pre>
-                  </figure>
-               </graphical-symbol>
-            </preferred>
-            <preferred id="_">
-               <letter-symbol>
-                  <name>
-                     <stem type="MathML">
-                        <math xmlns="http://www.w3.org/1998/Math/MathML">
-                           <msub>
-                              <mrow>
-                                 <mi>x</mi>
-                              </mrow>
-                              <mrow>
-                                 <mn>1</mn>
-                              </mrow>
-                           </msub>
-                        </math>
-                     </stem>
-                  </name>
-               </letter-symbol>
-            </preferred>
-            <fmt-preferred>
-               <p>
-                  <semx element="preferred" source="_">
-                     <strong/>
-                  </semx>
-                  ;
-                  <semx element="preferred" source="_">
-                     <strong>
-                        <stem type="MathML" id="_">
-                           <math xmlns="http://www.w3.org/1998/Math/MathML">
-                              <msub>
-                                 <mrow>
-                                    <mi>x</mi>
-                                 </mrow>
-                                 <mrow>
-                                    <mn>1</mn>
-                                 </mrow>
-                              </msub>
-                           </math>
-                        </stem>
-                        <fmt-stem type="MathML">
-                           <semx element="stem" source="_">
-                              <math xmlns="http://www.w3.org/1998/Math/MathML">
-                                 <mstyle mathvariant="bold">
-                                    <msub>
-                                       <mrow>
-                                          <mi>x</mi>
-                                       </mrow>
-                                       <mrow>
-                                          <mn>1</mn>
-                                       </mrow>
-                                    </msub>
-                                 </mstyle>
-                              </math>
-                              <asciimath>mathbf(x_(1))</asciimath>
-                           </semx>
-                        </fmt-stem>
-                     </strong>
-                  </semx>
-               </p>
-               <p>
-                  <semx element="preferred" source="_">
-                     <figure id="_" autonum="1">
-                        <fmt-name id="_">
-                           <span class="fmt-caption-label">
-                              <span class="fmt-element-name">Figure</span>
-                              <semx element="autonum" source="_">1</semx>
-                           </span>
-                        </fmt-name>
-                        <fmt-xref-label>
-                           <span class="fmt-element-name">Figure</span>
-                           <semx element="autonum" source="_">1</semx>
-                        </fmt-xref-label>
-                        <pre id="_">&lt;LITERAL&gt; FIGURATIVE</pre>
-                     </figure>
-                  </semx>
-               </p>
-            </fmt-preferred>
-            <domain id="_">rice</domain>
-            <definition id="_">
-               <verbal-definition>
-                  <p original-id="_">rice retaining its husk after threshing</p>
-               </verbal-definition>
-            </definition>
-            <fmt-definition id="_">
-               <semx element="definition" source="_">
-                  <p id="_">
-                     &lt;
-                     <semx element="domain" source="_">rice</semx>
-                     &gt; rice retaining its husk after threshing
-                  </p>
-               </semx>
-            </fmt-definition>
-         </term>
-      </terms>
-    PRESXML
-    expect(strip_guid(Canon.format_xml(Nokogiri::XML(IsoDoc::PresentationXMLConvert
-      .new(presxml_options)
-       .convert("test", input, true))
-      .at("//xmlns:terms").to_xml)))
-      .to be_equivalent_to Canon.format_xml(presxml)
-  end
-
   it "processes IsoXML term with nonverbal definitions" do
     input = <<~INPUT
           <iso-standard xmlns="http://riboseinc.com/isoxml">
@@ -2019,220 +1441,6 @@ RSpec.describe IsoDoc do
       .to be_equivalent_to Canon.format_xml(presxml)
   end
 
-  it "processes related terms" do
-    input = <<~INPUT
-      <iso-standard xmlns="http://riboseinc.com/isoxml">
-      <sections>
-      <terms id='A' obligation='normative'>
-            <title>Terms and definitions</title>
-            <term id='second'>
-        <preferred>
-          <expression>
-            <name>Second Term</name>
-          </expression>
-        <field-of-application>Field</field-of-application>
-        <usage-info>Usage Info 1</usage-info>
-        </preferred>
-        <definition><verbal-definition>Definition 1</verbal-definition></definition>
-      </term>
-      <term id="C">
-      <preferred language='fr' script='Latn' type='prefix'>
-                <expression>
-                  <name>First Designation</name>
-                  </expression></preferred>
-        <related type='contrast'>
-          <preferred>
-            <expression>
-              <name>Fifth Designation</name>
-              <grammar>
-                <gender>neuter</gender>
-              </grammar>
-            </expression>
-          </preferred>
-          <xref target='second'/>
-        </related>
-        <related type='see'>
-          <preferred>
-            <expression>
-              <name>Fifth Designation</name>
-              <grammar>
-                <gender>neuter</gender>
-              </grammar>
-            </expression>
-          </preferred>
-        </related>
-        <related type='seealso'>
-          <xref target='second'/>
-        </related>
-        <definition><verbal-definition>Definition 2</verbal-definition></definition>
-      </term>
-          </terms>
-        </sections>
-      </iso-standard>
-    INPUT
-    output = <<~OUTPUT
-      <terms id="A" obligation="normative" displayorder="2">
-          <title id="_">Terms and definitions</title>
-          <fmt-title id="_" depth="1">
-             <span class="fmt-caption-label">
-                <semx element="autonum" source="A">1</semx>
-                <span class="fmt-autonum-delim">.</span>
-             </span>
-             <span class="fmt-caption-delim">
-                <tab/>
-             </span>
-             <semx element="title" source="_">Terms and definitions</semx>
-          </fmt-title>
-          <fmt-xref-label>
-             <span class="fmt-element-name">Clause</span>
-             <semx element="autonum" source="A">1</semx>
-          </fmt-xref-label>
-          <term id="second">
-             <fmt-name id="_">
-                <span class="fmt-caption-label">
-                   <semx element="autonum" source="A">1</semx>
-                   <span class="fmt-autonum-delim">.</span>
-                   <semx element="autonum" source="second">1</semx>
-                   <span class="fmt-autonum-delim">.</span>
-                </span>
-             </fmt-name>
-             <fmt-xref-label>
-                <span class="fmt-element-name">Clause</span>
-                <semx element="autonum" source="A">1</semx>
-                <span class="fmt-autonum-delim">.</span>
-                <semx element="autonum" source="second">1</semx>
-             </fmt-xref-label>
-             <preferred id="_">
-                <expression>
-                   <name>Second Term</name>
-                </expression>
-                <field-of-application id="_">Field</field-of-application>
-                <usage-info id="_">Usage Info 1</usage-info>
-             </preferred>
-             <fmt-preferred>
-                <p>
-                   <semx element="preferred" source="_">
-                      <strong>Second Term</strong>
-                      <span class="fmt-designation-field">
-                         , &lt;
-                         <semx element="field-of-application" source="_">Field</semx>
-                         ,
-                         <semx element="usage-info" source="_">Usage Info 1</semx>
-                         &gt;
-                      </span>
-                   </semx>
-                </p>
-             </fmt-preferred>
-             <definition id="_">
-                <verbal-definition>Definition 1</verbal-definition>
-             </definition>
-             <fmt-definition id="_">
-                <semx element="definition" source="_">Definition 1</semx>
-             </fmt-definition>
-          </term>
-          <term id="C">
-             <fmt-name id="_">
-                <span class="fmt-caption-label">
-                   <semx element="autonum" source="A">1</semx>
-                   <span class="fmt-autonum-delim">.</span>
-                   <semx element="autonum" source="C">2</semx>
-                   <span class="fmt-autonum-delim">.</span>
-                </span>
-             </fmt-name>
-             <fmt-xref-label>
-                <span class="fmt-element-name">Clause</span>
-                <semx element="autonum" source="A">1</semx>
-                <span class="fmt-autonum-delim">.</span>
-                <semx element="autonum" source="C">2</semx>
-             </fmt-xref-label>
-             <preferred language="fr" script="Latn" type="prefix" id="_">
-                <expression>
-                   <name>First Designation</name>
-                </expression>
-             </preferred>
-             <fmt-preferred>
-                <p>
-                   <semx element="preferred" source="_">
-                      <strong>First Designation</strong>
-                   </semx>
-                </p>
-             </fmt-preferred>
-             <related type="contrast" id="_">
-                <preferred id="_">
-                   <expression>
-                      <name>Fifth Designation</name>
-                      <grammar>
-                         <gender>neuter</gender>
-                      </grammar>
-                   </expression>
-                </preferred>
-                <xref target="second"/>
-             </related>
-             <related type="see" id="_">
-                <preferred id="_">
-                   <expression>
-                      <name>Fifth Designation</name>
-                      <grammar>
-                         <gender>neuter</gender>
-                      </grammar>
-                   </expression>
-                </preferred>
-             </related>
-             <related type="seealso" id="_">
-                <xref target="second"/>
-             </related>
-             <fmt-related>
-                <semx element="related" source="_">
-                   <p>
-                      <strong>CONTRAST:</strong>
-                      <em>
-                            <semx element="preferred" source="_">
-                               <strong>Fifth Designation</strong>
-                               , n
-                            </semx>
-                      </em>
-                      (
-               <xref target="second" id="_"/>
-               <semx element="xref" source="_">
-                  <fmt-xref target="second">
-                     <span class="fmt-element-name">Clause</span>
-                     <semx element="autonum" source="A">1</semx>
-                     <span class="fmt-autonum-delim">.</span>
-                     <semx element="autonum" source="second">1</semx>
-                  </fmt-xref>
-               </semx>
-                      )
-                   </p>
-                </semx>
-                <semx element="related" source="_">
-                   <p>
-                      <strong>SEE:</strong>
-                      <strong>**RELATED TERM NOT FOUND**</strong>
-                   </p>
-                </semx>
-                <semx element="related" source="_">
-                   <p>
-                      <strong>SEE ALSO:</strong>
-                      <strong>**RELATED TERM NOT FOUND**</strong>
-                   </p>
-                </semx>
-             </fmt-related>
-             <definition id="_">
-                <verbal-definition>Definition 2</verbal-definition>
-             </definition>
-             <fmt-definition id="_">
-                <semx element="definition" source="_">Definition 2</semx>
-             </fmt-definition>
-          </term>
-       </terms>
-    OUTPUT
-    expect(strip_guid(Canon.format_xml(Nokogiri::XML(IsoDoc::PresentationXMLConvert
-      .new(presxml_options)
-       .convert("test", input, true))
-      .at("//xmlns:terms").to_xml)))
-      .to be_equivalent_to Canon.format_xml(output)
-  end
-
   it "processes IsoXML term with different term source statuses" do
     input = <<~INPUT
           <iso-standard xmlns="http://riboseinc.com/isoxml">
@@ -2297,178 +1505,178 @@ RSpec.describe IsoDoc do
       </sections></iso-standard>
     INPUT
     output = <<~OUTPUT
-       <terms id="terms_and_definitions" obligation="normative" displayorder="2">
-          <title id="_">Terms and Definitions</title>
-          <fmt-title id="_" depth="1">
-             <span class="fmt-caption-label">
-                <semx element="autonum" source="terms_and_definitions">1</semx>
-                <span class="fmt-autonum-delim">.</span>
-             </span>
-             <span class="fmt-caption-delim">
-                <tab/>
-             </span>
-             <semx element="title" source="_">Terms and Definitions</semx>
-          </fmt-title>
-          <fmt-xref-label>
-             <span class="fmt-element-name">Clause</span>
-             <semx element="autonum" source="terms_and_definitions">1</semx>
-          </fmt-xref-label>
-          <p>For the purposes of this document, the following terms and definitions apply.</p>
-          <term id="paddy1">
-             <fmt-name id="_">
-                <span class="fmt-caption-label">
-                   <semx element="autonum" source="terms_and_definitions">1</semx>
-                   <span class="fmt-autonum-delim">.</span>
-                   <semx element="autonum" source="paddy1">1</semx>
-                   <span class="fmt-autonum-delim">.</span>
-                </span>
-             </fmt-name>
-             <fmt-xref-label>
-                <span class="fmt-element-name">Clause</span>
-                <semx element="autonum" source="terms_and_definitions">1</semx>
-                <span class="fmt-autonum-delim">.</span>
-                <semx element="autonum" source="paddy1">1</semx>
-             </fmt-xref-label>
-             <preferred id="_">
-                <expression>
-                   <name>paddy</name>
-                </expression>
-             </preferred>
-             <fmt-preferred>
-                <p>
-                   <semx element="preferred" source="_">
-                      <strong>paddy</strong>
-                   </semx>
-                </p>
-             </fmt-preferred>
-             <definition id="_">
-                <verbal-definition>
-                   <p original-id="_">rice retaining its husk after threshing</p>
-                </verbal-definition>
-             </definition>
-             <fmt-definition id="_">
-                <semx element="definition" source="_">
-                   <p id="_">rice retaining its husk after threshing</p>
-                </semx>
-             </fmt-definition>
-             <source status="identical" id="_">
-                <origin citeas="">
-                   <termref base="IEV" target="xyz">t1</termref>
-                </origin>
-                <modification id="_">
-                   <p id="_">with adjustments</p>
-                </modification>
-             </source>
-             <source status="adapted" id="_">
-                <origin citeas="">
-                   <termref base="IEV" target="xyz"/>
-                </origin>
-                <modification id="_">
-                   <p original-id="_">with adjustments</p>
-                </modification>
-             </source>
-             <source status="modified" id="_">
-                <origin citeas="">
-                   <termref base="IEV" target="xyz"/>
-                </origin>
-                <modification id="_">
-                   <p original-id="_">with adjustments</p>
-                </modification>
-             </source>
-             <source status="identical" id="_">
-                <origin citeas="">
-                   <termref base="IEV" target="xyz">t1</termref>
-                </origin>
-             </source>
-             <source status="adapted" id="_">
-                <origin citeas="">
-                   <termref base="IEV" target="xyz"/>
-                </origin>
-             </source>
-             <source status="modified" id="_">
-                <origin citeas="">
-                   <termref base="IEV" target="xyz"/>
-                </origin>
-             </source>
-             <fmt-termsource status="identical">
-                [SOURCE:
-                <semx element="source" source="_">
-                   <origin citeas="" id="_">
-                      <termref base="IEV" target="xyz">t1</termref>
-                   </origin>
-                   <semx element="origin" source="_">
-                      <fmt-origin citeas="">
-                         <termref base="IEV" target="xyz">t1</termref>
-                      </fmt-origin>
-                   </semx>
-                   —
-                   <semx element="modification" source="_">with adjustments</semx>
-                </semx>
-                ;
-                <semx element="source" source="_">
-                   <origin citeas="" id="_">
-                      <termref base="IEV" target="xyz"/>
-                   </origin>
-                   <semx element="origin" source="_">
-                      <fmt-origin citeas="">
-                         <termref base="IEV" target="xyz"/>
-                      </fmt-origin>
-                   </semx>
-                   , adapted —
-                   <semx element="modification" source="_">with adjustments</semx>
-                </semx>
-                ;
-                <semx element="source" source="_">
-                   <origin citeas="" id="_">
-                      <termref base="IEV" target="xyz"/>
-                   </origin>
-                   <semx element="origin" source="_">
-                      <fmt-origin citeas="">
-                         <termref base="IEV" target="xyz"/>
-                      </fmt-origin>
-                   </semx>
-                   , modified —
-                   <semx element="modification" source="_">with adjustments</semx>
-                </semx>
-                ;
-                <semx element="source" source="_">
-                   <origin citeas="" id="_">
-                      <termref base="IEV" target="xyz">t1</termref>
-                   </origin>
-                   <semx element="origin" source="_">
-                      <fmt-origin citeas="">
-                         <termref base="IEV" target="xyz">t1</termref>
-                      </fmt-origin>
-                   </semx>
-                </semx>
-                ;
-                <semx element="source" source="_">
-                   <origin citeas="" id="_">
-                      <termref base="IEV" target="xyz"/>
-                   </origin>
-                   <semx element="origin" source="_">
-                      <fmt-origin citeas="">
-                         <termref base="IEV" target="xyz"/>
-                      </fmt-origin>
-                   </semx>
-                   , adapted
-                </semx>
-                ;
-                <semx element="source" source="_">
-                   <origin citeas="" id="_">
-                      <termref base="IEV" target="xyz"/>
-                   </origin>
-                   <semx element="origin" source="_">
-                      <fmt-origin citeas="">
-                         <termref base="IEV" target="xyz"/>
-                      </fmt-origin>
-                   </semx>
-                   , modified
-                </semx>
-                ]
-             </fmt-termsource>
-          </term>
-       </terms>
+      <terms id="terms_and_definitions" obligation="normative" displayorder="2">
+         <title id="_">Terms and Definitions</title>
+         <fmt-title id="_" depth="1">
+            <span class="fmt-caption-label">
+               <semx element="autonum" source="terms_and_definitions">1</semx>
+               <span class="fmt-autonum-delim">.</span>
+            </span>
+            <span class="fmt-caption-delim">
+               <tab/>
+            </span>
+            <semx element="title" source="_">Terms and Definitions</semx>
+         </fmt-title>
+         <fmt-xref-label>
+            <span class="fmt-element-name">Clause</span>
+            <semx element="autonum" source="terms_and_definitions">1</semx>
+         </fmt-xref-label>
+         <p>For the purposes of this document, the following terms and definitions apply.</p>
+         <term id="paddy1">
+            <fmt-name id="_">
+               <span class="fmt-caption-label">
+                  <semx element="autonum" source="terms_and_definitions">1</semx>
+                  <span class="fmt-autonum-delim">.</span>
+                  <semx element="autonum" source="paddy1">1</semx>
+                  <span class="fmt-autonum-delim">.</span>
+               </span>
+            </fmt-name>
+            <fmt-xref-label>
+               <span class="fmt-element-name">Clause</span>
+               <semx element="autonum" source="terms_and_definitions">1</semx>
+               <span class="fmt-autonum-delim">.</span>
+               <semx element="autonum" source="paddy1">1</semx>
+            </fmt-xref-label>
+            <preferred id="_">
+               <expression>
+                  <name>paddy</name>
+               </expression>
+            </preferred>
+            <fmt-preferred>
+               <p>
+                  <semx element="preferred" source="_">
+                     <strong>paddy</strong>
+                  </semx>
+               </p>
+            </fmt-preferred>
+            <definition id="_">
+               <verbal-definition>
+                  <p original-id="_">rice retaining its husk after threshing</p>
+               </verbal-definition>
+            </definition>
+            <fmt-definition id="_">
+               <semx element="definition" source="_">
+                  <p id="_">rice retaining its husk after threshing</p>
+               </semx>
+            </fmt-definition>
+            <source status="identical" id="_">
+               <origin citeas="">
+                  <termref base="IEV" target="xyz">t1</termref>
+               </origin>
+               <modification id="_">
+                  <p id="_">with adjustments</p>
+               </modification>
+            </source>
+            <source status="adapted" id="_">
+               <origin citeas="">
+                  <termref base="IEV" target="xyz"/>
+               </origin>
+               <modification id="_">
+                  <p original-id="_">with adjustments</p>
+               </modification>
+            </source>
+            <source status="modified" id="_">
+               <origin citeas="">
+                  <termref base="IEV" target="xyz"/>
+               </origin>
+               <modification id="_">
+                  <p original-id="_">with adjustments</p>
+               </modification>
+            </source>
+            <source status="identical" id="_">
+               <origin citeas="">
+                  <termref base="IEV" target="xyz">t1</termref>
+               </origin>
+            </source>
+            <source status="adapted" id="_">
+               <origin citeas="">
+                  <termref base="IEV" target="xyz"/>
+               </origin>
+            </source>
+            <source status="modified" id="_">
+               <origin citeas="">
+                  <termref base="IEV" target="xyz"/>
+               </origin>
+            </source>
+            <fmt-termsource status="identical">
+               [SOURCE:
+               <semx element="source" source="_">
+                  <origin citeas="" id="_">
+                     <termref base="IEV" target="xyz">t1</termref>
+                  </origin>
+                  <semx element="origin" source="_">
+                     <fmt-origin citeas="">
+                        <termref base="IEV" target="xyz">t1</termref>
+                     </fmt-origin>
+                  </semx>
+                  —
+                  <semx element="modification" source="_">with adjustments</semx>
+               </semx>
+               ;
+               <semx element="source" source="_">
+                  <origin citeas="" id="_">
+                     <termref base="IEV" target="xyz"/>
+                  </origin>
+                  <semx element="origin" source="_">
+                     <fmt-origin citeas="">
+                        <termref base="IEV" target="xyz"/>
+                     </fmt-origin>
+                  </semx>
+                  , adapted —
+                  <semx element="modification" source="_">with adjustments</semx>
+               </semx>
+               ;
+               <semx element="source" source="_">
+                  <origin citeas="" id="_">
+                     <termref base="IEV" target="xyz"/>
+                  </origin>
+                  <semx element="origin" source="_">
+                     <fmt-origin citeas="">
+                        <termref base="IEV" target="xyz"/>
+                     </fmt-origin>
+                  </semx>
+                  , modified —
+                  <semx element="modification" source="_">with adjustments</semx>
+               </semx>
+               ;
+               <semx element="source" source="_">
+                  <origin citeas="" id="_">
+                     <termref base="IEV" target="xyz">t1</termref>
+                  </origin>
+                  <semx element="origin" source="_">
+                     <fmt-origin citeas="">
+                        <termref base="IEV" target="xyz">t1</termref>
+                     </fmt-origin>
+                  </semx>
+               </semx>
+               ;
+               <semx element="source" source="_">
+                  <origin citeas="" id="_">
+                     <termref base="IEV" target="xyz"/>
+                  </origin>
+                  <semx element="origin" source="_">
+                     <fmt-origin citeas="">
+                        <termref base="IEV" target="xyz"/>
+                     </fmt-origin>
+                  </semx>
+                  , adapted
+               </semx>
+               ;
+               <semx element="source" source="_">
+                  <origin citeas="" id="_">
+                     <termref base="IEV" target="xyz"/>
+                  </origin>
+                  <semx element="origin" source="_">
+                     <fmt-origin citeas="">
+                        <termref base="IEV" target="xyz"/>
+                     </fmt-origin>
+                  </semx>
+                  , modified
+               </semx>
+               ]
+            </fmt-termsource>
+         </term>
+      </terms>
     OUTPUT
     expect(strip_guid(Canon.format_xml(Nokogiri::XML(IsoDoc::PresentationXMLConvert
           .new(presxml_options)
@@ -2476,178 +1684,178 @@ RSpec.describe IsoDoc do
           .at("//xmlns:terms").to_xml)))
       .to be_equivalent_to Canon.format_xml(output)
     output = <<~OUTPUT
-       <terms id="terms_and_definitions" obligation="normative" displayorder="2">
-          <title id="_">Terms and Definitions</title>
-          <fmt-title id="_" depth="1">
-             <span class="fmt-caption-label">
-                <semx element="autonum" source="terms_and_definitions">1</semx>
-                <span class="fmt-autonum-delim">.</span>
-             </span>
-             <span class="fmt-caption-delim">
-                <tab/>
-             </span>
-             <semx element="title" source="_">Terms and Definitions</semx>
-          </fmt-title>
-          <fmt-xref-label>
-             <span class="fmt-element-name">Klausel</span>
-             <semx element="autonum" source="terms_and_definitions">1</semx>
-          </fmt-xref-label>
-          <p>For the purposes of this document, the following terms and definitions apply.</p>
-          <term id="paddy1">
-             <fmt-name id="_">
-                <span class="fmt-caption-label">
-                   <semx element="autonum" source="terms_and_definitions">1</semx>
-                   <span class="fmt-autonum-delim">.</span>
-                   <semx element="autonum" source="paddy1">1</semx>
-                   <span class="fmt-autonum-delim">.</span>
-                </span>
-             </fmt-name>
-             <fmt-xref-label>
-                <span class="fmt-element-name">Klausel</span>
-                <semx element="autonum" source="terms_and_definitions">1</semx>
-                <span class="fmt-autonum-delim">.</span>
-                <semx element="autonum" source="paddy1">1</semx>
-             </fmt-xref-label>
-             <preferred id="_">
-                <expression>
-                   <name>paddy</name>
-                </expression>
-             </preferred>
-             <fmt-preferred>
-                <p>
-                   <semx element="preferred" source="_">
-                      <strong>paddy</strong>
-                   </semx>
-                </p>
-             </fmt-preferred>
-             <definition id="_">
-                <verbal-definition>
-                   <p original-id="_">rice retaining its husk after threshing</p>
-                </verbal-definition>
-             </definition>
-             <fmt-definition id="_">
-                <semx element="definition" source="_">
-                   <p id="_">rice retaining its husk after threshing</p>
-                </semx>
-             </fmt-definition>
-             <source status="identical" id="_">
-                <origin citeas="">
-                   <termref base="IEV" target="xyz">t1</termref>
-                </origin>
-                <modification id="_">
-                   <p id="_">with adjustments</p>
-                </modification>
-             </source>
-             <source status="adapted" id="_">
-                <origin citeas="">
-                   <termref base="IEV" target="xyz"/>
-                </origin>
-                <modification id="_">
-                   <p original-id="_">with adjustments</p>
-                </modification>
-             </source>
-             <source status="modified" id="_">
-                <origin citeas="">
-                   <termref base="IEV" target="xyz"/>
-                </origin>
-                <modification id="_">
-                   <p original-id="_">with adjustments</p>
-                </modification>
-             </source>
-             <source status="identical" id="_">
-                <origin citeas="">
-                   <termref base="IEV" target="xyz">t1</termref>
-                </origin>
-             </source>
-             <source status="adapted" id="_">
-                <origin citeas="">
-                   <termref base="IEV" target="xyz"/>
-                </origin>
-             </source>
-             <source status="modified" id="_">
-                <origin citeas="">
-                   <termref base="IEV" target="xyz"/>
-                </origin>
-             </source>
-             <fmt-termsource status="identical">
-                [QUELLE:
-                <semx element="source" source="_">
-                   <origin citeas="" id="_">
-                      <termref base="IEV" target="xyz">t1</termref>
-                   </origin>
-                   <semx element="origin" source="_">
-                      <fmt-origin citeas="">
-                         <termref base="IEV" target="xyz">t1</termref>
-                      </fmt-origin>
-                   </semx>
-                   —
-                   <semx element="modification" source="_">with adjustments</semx>
-                </semx>
-                ;
-                <semx element="source" source="_">
-                   <origin citeas="" id="_">
-                      <termref base="IEV" target="xyz"/>
-                   </origin>
-                   <semx element="origin" source="_">
-                      <fmt-origin citeas="">
-                         <termref base="IEV" target="xyz"/>
-                      </fmt-origin>
-                   </semx>
-                   , angepasst —
-                   <semx element="modification" source="_">with adjustments</semx>
-                </semx>
-                ;
-                <semx element="source" source="_">
-                   <origin citeas="" id="_">
-                      <termref base="IEV" target="xyz"/>
-                   </origin>
-                   <semx element="origin" source="_">
-                      <fmt-origin citeas="">
-                         <termref base="IEV" target="xyz"/>
-                      </fmt-origin>
-                   </semx>
-                   , geändert —
-                   <semx element="modification" source="_">with adjustments</semx>
-                </semx>
-                ;
-                <semx element="source" source="_">
-                   <origin citeas="" id="_">
-                      <termref base="IEV" target="xyz">t1</termref>
-                   </origin>
-                   <semx element="origin" source="_">
-                      <fmt-origin citeas="">
-                         <termref base="IEV" target="xyz">t1</termref>
-                      </fmt-origin>
-                   </semx>
-                </semx>
-                ;
-                <semx element="source" source="_">
-                   <origin citeas="" id="_">
-                      <termref base="IEV" target="xyz"/>
-                   </origin>
-                   <semx element="origin" source="_">
-                      <fmt-origin citeas="">
-                         <termref base="IEV" target="xyz"/>
-                      </fmt-origin>
-                   </semx>
-                   , angepasst
-                </semx>
-                ;
-                <semx element="source" source="_">
-                   <origin citeas="" id="_">
-                      <termref base="IEV" target="xyz"/>
-                   </origin>
-                   <semx element="origin" source="_">
-                      <fmt-origin citeas="">
-                         <termref base="IEV" target="xyz"/>
-                      </fmt-origin>
-                   </semx>
-                   , geändert
-                </semx>
-                ]
-             </fmt-termsource>
-          </term>
-       </terms>
+      <terms id="terms_and_definitions" obligation="normative" displayorder="2">
+         <title id="_">Terms and Definitions</title>
+         <fmt-title id="_" depth="1">
+            <span class="fmt-caption-label">
+               <semx element="autonum" source="terms_and_definitions">1</semx>
+               <span class="fmt-autonum-delim">.</span>
+            </span>
+            <span class="fmt-caption-delim">
+               <tab/>
+            </span>
+            <semx element="title" source="_">Terms and Definitions</semx>
+         </fmt-title>
+         <fmt-xref-label>
+            <span class="fmt-element-name">Klausel</span>
+            <semx element="autonum" source="terms_and_definitions">1</semx>
+         </fmt-xref-label>
+         <p>For the purposes of this document, the following terms and definitions apply.</p>
+         <term id="paddy1">
+            <fmt-name id="_">
+               <span class="fmt-caption-label">
+                  <semx element="autonum" source="terms_and_definitions">1</semx>
+                  <span class="fmt-autonum-delim">.</span>
+                  <semx element="autonum" source="paddy1">1</semx>
+                  <span class="fmt-autonum-delim">.</span>
+               </span>
+            </fmt-name>
+            <fmt-xref-label>
+               <span class="fmt-element-name">Klausel</span>
+               <semx element="autonum" source="terms_and_definitions">1</semx>
+               <span class="fmt-autonum-delim">.</span>
+               <semx element="autonum" source="paddy1">1</semx>
+            </fmt-xref-label>
+            <preferred id="_">
+               <expression>
+                  <name>paddy</name>
+               </expression>
+            </preferred>
+            <fmt-preferred>
+               <p>
+                  <semx element="preferred" source="_">
+                     <strong>paddy</strong>
+                  </semx>
+               </p>
+            </fmt-preferred>
+            <definition id="_">
+               <verbal-definition>
+                  <p original-id="_">rice retaining its husk after threshing</p>
+               </verbal-definition>
+            </definition>
+            <fmt-definition id="_">
+               <semx element="definition" source="_">
+                  <p id="_">rice retaining its husk after threshing</p>
+               </semx>
+            </fmt-definition>
+            <source status="identical" id="_">
+               <origin citeas="">
+                  <termref base="IEV" target="xyz">t1</termref>
+               </origin>
+               <modification id="_">
+                  <p id="_">with adjustments</p>
+               </modification>
+            </source>
+            <source status="adapted" id="_">
+               <origin citeas="">
+                  <termref base="IEV" target="xyz"/>
+               </origin>
+               <modification id="_">
+                  <p original-id="_">with adjustments</p>
+               </modification>
+            </source>
+            <source status="modified" id="_">
+               <origin citeas="">
+                  <termref base="IEV" target="xyz"/>
+               </origin>
+               <modification id="_">
+                  <p original-id="_">with adjustments</p>
+               </modification>
+            </source>
+            <source status="identical" id="_">
+               <origin citeas="">
+                  <termref base="IEV" target="xyz">t1</termref>
+               </origin>
+            </source>
+            <source status="adapted" id="_">
+               <origin citeas="">
+                  <termref base="IEV" target="xyz"/>
+               </origin>
+            </source>
+            <source status="modified" id="_">
+               <origin citeas="">
+                  <termref base="IEV" target="xyz"/>
+               </origin>
+            </source>
+            <fmt-termsource status="identical">
+               [QUELLE:
+               <semx element="source" source="_">
+                  <origin citeas="" id="_">
+                     <termref base="IEV" target="xyz">t1</termref>
+                  </origin>
+                  <semx element="origin" source="_">
+                     <fmt-origin citeas="">
+                        <termref base="IEV" target="xyz">t1</termref>
+                     </fmt-origin>
+                  </semx>
+                  —
+                  <semx element="modification" source="_">with adjustments</semx>
+               </semx>
+               ;
+               <semx element="source" source="_">
+                  <origin citeas="" id="_">
+                     <termref base="IEV" target="xyz"/>
+                  </origin>
+                  <semx element="origin" source="_">
+                     <fmt-origin citeas="">
+                        <termref base="IEV" target="xyz"/>
+                     </fmt-origin>
+                  </semx>
+                  , angepasst —
+                  <semx element="modification" source="_">with adjustments</semx>
+               </semx>
+               ;
+               <semx element="source" source="_">
+                  <origin citeas="" id="_">
+                     <termref base="IEV" target="xyz"/>
+                  </origin>
+                  <semx element="origin" source="_">
+                     <fmt-origin citeas="">
+                        <termref base="IEV" target="xyz"/>
+                     </fmt-origin>
+                  </semx>
+                  , geändert —
+                  <semx element="modification" source="_">with adjustments</semx>
+               </semx>
+               ;
+               <semx element="source" source="_">
+                  <origin citeas="" id="_">
+                     <termref base="IEV" target="xyz">t1</termref>
+                  </origin>
+                  <semx element="origin" source="_">
+                     <fmt-origin citeas="">
+                        <termref base="IEV" target="xyz">t1</termref>
+                     </fmt-origin>
+                  </semx>
+               </semx>
+               ;
+               <semx element="source" source="_">
+                  <origin citeas="" id="_">
+                     <termref base="IEV" target="xyz"/>
+                  </origin>
+                  <semx element="origin" source="_">
+                     <fmt-origin citeas="">
+                        <termref base="IEV" target="xyz"/>
+                     </fmt-origin>
+                  </semx>
+                  , angepasst
+               </semx>
+               ;
+               <semx element="source" source="_">
+                  <origin citeas="" id="_">
+                     <termref base="IEV" target="xyz"/>
+                  </origin>
+                  <semx element="origin" source="_">
+                     <fmt-origin citeas="">
+                        <termref base="IEV" target="xyz"/>
+                     </fmt-origin>
+                  </semx>
+                  , geändert
+               </semx>
+               ]
+            </fmt-termsource>
+         </term>
+      </terms>
     OUTPUT
     expect(strip_guid(Canon.format_xml(Nokogiri::XML(IsoDoc::PresentationXMLConvert
           .new(presxml_options)
